Better support of starting values and more stable PQL/MQL algorithm
`mclogit()` and `mblogit()` now have an infrastructure that allows
to provide starting values for models with random effects.
Starting values should be given for the coefficents and for
the variance parameters. If the PQL method is used, also
starting values for the random effects have to be provided.

The PQL/MQL algorithm is made more stable by re-using in for the inner iteration
the intermediate estimates of coefficients, variance parameters, and random
parameters from previous outer iterations.